backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-multipart==0.0.18
pillow==11.0.0
pillow-heif==0.20.0
